requirements.txt
numpy
pandas
requests
python-binance
loguru
pytest
python-dotenv